Log errors on local deploy (#184)
* Allow more information to be passed back from goal execution

and rename a file

* remove duplication: local jar deploy, use general

* Use failure description when failed, local deploy:

* eliminate source deployer, 1/n: pass branch in DeployParams

* Pass branch to targeter

* 2/n: calculate managed deployment key in targeting

* 3/n: pass TargetInfo to undeploy

* n/n be a deployer instead of a special thing

* n+1, delete method

* n+2, make imageName optional

specify this on the artifactStore. This also fixes a bug of using the global one all the time

* incorporate image check in error handling, 1/p: use generic retry

This is going to enable retry commands for every goal, not just deploy

* 2/p: move error reporting and logging out a layer

* Fix: do not trigger on every pending status lol

* Interpret any log that's short

* p: fix the image URL being optional thing

it's all set up for it now. Next error please

* Fix: set undefined after we access it

* Fix: might be undefined, don't kill it then

* Move three files

put the generic things in a generic place

* remove a circular dependency between approvalGate and Goal

* Explicitly depend on dependencies

* Do not reference index from spi or common

* do not reference index from handlers

* refine SourceDeployer

* Rename Deployer to ArtifactDeployer

* restore SourceDeployer 1/m: extract loggy bit

* make the maven one a sourceDeployer again

* fix: moved a file, must change relative path

* Lint

* Polishing: small renaming and Typedoc

+++ src/handlers/events/delivery/deploy/local/appManagement.ts
@@ -1,6 +1,8 @@
 import { logger } from "@atomist/automation-client";
 import { RemoteRepoRef } from "@atomist/automation-client/operations/common/RepoId";
 import { ChildProcess } from "child_process";
+import { TargetInfo } from "../../../../../spi/deploy/Deployment";
+import { Targeter } from "../deploy";
 
 export interface BranchRepoRef extends RemoteRepoRef {
     branch?: string;
@@ -8,6 +10,30 @@ export interface BranchRepoRef extends RemoteRepoRef {
 
 export function isBranchRepoRef(rr: RemoteRepoRef): rr is BranchRepoRef {
     return !!(rr as BranchRepoRef).branch;
+}
+
+export interface ManagedDeploymentTargetInfo extends TargetInfo {
+    managedDeploymentKey: BranchRepoRef;
+}
+
+export const ManagedDeploymentTargeter: Targeter<ManagedDeploymentTargetInfo> = (id: RemoteRepoRef, branch: string) => {
+    const branchId = {...id, branch};
+    return {
+        name: "Run alongside this automation",
+        description: `Locally run ${id.sha} from branch ${branch}`,
+        managedDeploymentKey: branchId,
+    };
+};
+
+// this is currently used in shutdown
+// because Superseded should be per-branch, but isn't yet.
+// At least this makes it explicit we don't have it quite right yet
+export function targetInfoForAllBranches(id: RemoteRepoRef): ManagedDeploymentTargetInfo {
+    return {
+        managedDeploymentKey: {...id, branch: undefined},
+        name: "Run alongside this automation",
+        description: `Locally run ${id.sha} from an unknown branch`,
+    };
 }
 
 /**
@@ -62,12 +88,12 @@ export class ManagedDeployments {
     public async terminateIfRunning(id: BranchRepoRef): Promise<any> {
         const victim = this.deployments.find(d => d.id.sha === id.sha ||
             (d.id.owner === id.owner && d.id.repo === id.repo && !!id.branch && d.id.branch === id.branch));
-        if (!!victim) {
+        if (!!victim && !!victim.childProcess) {
             victim.childProcess.kill();
             // Keep the port but deallocate the process
-            victim.childProcess = undefined;
             logger.info("Killed app [%j] with pid %d, but continuing to reserve port [%d]",
                 id, victim.childProcess.pid, victim.port);
+            victim.childProcess = undefined;
         } else {
             logger.info("Was asked to kill app [%j], but no eligible process found", id);
         }
